% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_format.R
\name{as_phenology}
\alias{as_phenology}
\title{Puts data in right format for a MoniTool model run}
\usage{
as_phenology(
  data,
  reference.date,
  max.days = 1,
  min.obs = 10,
  sites.together = F
)
}
\arguments{
\item{data}{is a dataframe which must have a column called 'date' which was the date on which
the count was done, along with at least one of columns entitled'nests' or 'activities' (can be both).
May also optionally have columns called 'beach' to differentiate sites, and 'season' to name nesting
seasons. If 'season' is missing it will be guessed from the dates provided.}

\item{max.days}{is a number giving the maximum number of days over which the
activities counted in each survey may have occurred. So \code{max.days = 1} means all activities occurred on the
the night before survey was done (for morning counts). For \code{max.days > 1} counted
activities are assumed to have occurred over all days since the previous count, up to
a maximum of \code{date - max.days}. This requires you to make some informed assumption about
how many days activities remain visible for. For more complex monitoring designs it is
also possible to include a column in \code{data} called \code{datestart} which contains the earliest dates
when counts in each survey could have occurred. For example, if tracks were obliterated two days
before the survey on some occassions, then \code{datestart} could be set to
\verb{date - 2 days} for these observations and to \code{date - max.days} for the remainder.}

\item{min.obs}{is a single numeric giving the minimum number of observations that are needed
for a given beach in a given season to fit a model. Beaches with \verb{< min.obs} counts will
be removed from the data}

\item{sites.together}{A logical indicating whether counts on beaches in each season should be modelled together.
This is intended to allow users to estimate some joint parameters across sites (e.g. peak date) but the
initialization process is currently too slow with multiple beaches so it is recommended to leave sites.together = F.}
}
\description{
Puts data in right format for a MoniTool model run
}
