% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SbinomNim.R
\name{dSnbinomNim}
\alias{dSnbinomNim}
\title{This is a custom probability distribution giving the density for the sum of mutiple
random variables with negative binomial distributions (e.g. in this case, numbers of activities over multiple days).
The function is a \code{Nimble} implementation of the function in package \code{HelpersMG}.}
\usage{
dSnbinomNim(x, size, mu, nt, log = 0)
}
\arguments{
\item{x, size, mu, nt, log}{Definitions as given in HelpersMG::dSnbinom}
}
\description{
This is a custom probability distribution giving the density for the sum of mutiple
random variables with negative binomial distributions (e.g. in this case, numbers of activities over multiple days).
The function is a \code{Nimble} implementation of the function in package \code{HelpersMG}.
}
