% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_predict.R
\name{MT_annual_proportions}
\alias{MT_annual_proportions}
\title{We could optionally here allow use of the totals including observations (Y) or only predicted values (y.pred) to
assess changes in confidence intervals.}
\usage{
MT_annual_proportions(MTpred, interval = 0.95, full.posterior = F)
}
\description{
We could optionally here allow use of the totals including observations (Y) or only predicted values (y.pred) to
assess changes in confidence intervals.
}
