% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_phenology.R
\name{MT_fit}
\alias{MT_fit}
\title{This function currently requires that you specify a peak for initialising the model. It would be better it it just self
initialised. We could just use highest value in data or from another model e.g. GAM or maximum likelihood fit of this model.}
\usage{
MT_fit(
  data,
  nchains = 2,
  niter = 40000,
  nburnin = 10000,
  thin = 1,
  params = c("alpha", "s1", "s2", "tf", "tp", "phi"),
  init.attempts = 100,
  init.control = list(tp.min = 0, tp.max = 365, smin = 0.01, smax = 1),
  ncores = 1
)
}
\description{
This function currently requires that you specify a peak for initialising the model. It would be better it it just self
initialised. We could just use highest value in data or from another model e.g. GAM or maximum likelihood fit of this model.
}
