% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_phenology.R
\name{phenology.girondot}
\alias{phenology.girondot}
\title{Returns a numbleFunction implementing the phenology model of
Girondot et al. 2010 Estimating density of animals during migratory
waves: a new model applied to marine turtles at nesting sites.}
\usage{
phenology.girondot(t, E, B, Max, MinE, MinB, tf, tp)
}
\description{
Returns a numbleFunction implementing the phenology model of
Girondot et al. 2010 Estimating density of animals during migratory
waves: a new model applied to marine turtles at nesting sites.
}
