% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_phenology.R
\name{phenology.omeyer}
\alias{phenology.omeyer}
\title{Returns a numbleFunction implementing the Omeyer et al. 2022 model of mean number of counts
as a function of days.}
\usage{
phenology.omeyer(t, alpha, s1, s2, tf, tp)
}
\description{
Returns a numbleFunction implementing the Omeyer et al. 2022 model of mean number of counts
as a function of days.
}
