% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_format.R
\name{set_ref_date}
\alias{set_ref_date}
\title{Used internally by \code{MT_prep} to adjust dates using a reference date.}
\usage{
set_ref_date(x, ref_date)
}
\arguments{
\item{x}{A vector of class \code{Date} to be adjusted}

\item{ref_date}{A character vector giving the reference date in format \%d\%B or \%B\%d (e.g. '1 November', 'Dec 31')}
}
\description{
Used internally by \code{MT_prep} to adjust dates using a reference date.
}
